refactor: 수량 관련 컬럼 타입 변경 (Integer -> Long)

+++ src/main/java/kr/hhplus/be/server/application/order/dto/OrderCriteria.java
@@ -14,7 +14,7 @@ public record OrderCriteria() {
     
     public record OrderItem(
             Long productOptionId,
-            Integer quantity
+            Long quantity
     ) {}
