+++ src/main/java/kr/hhplus/be/server/interfaces/product/dto/ProductResponse.java
@@ -36,7 +36,7 @@ public record ProductResponse() {
             Long optionId,
             String optionValue,
             Long price,
-            Integer stock
+            Long stock
     ) {
         public static Option from(ProductResult.Option option) {
             return new Option(
